Replace random story Bob
It was 94 fahrenheit outside, so Father Christmas went for a walk. When they got to the White House, they stared in horror for a few moments, then turned into a slug and crawled away. BOb saw the whole thing, but was not surprised — Father Christmas weighs 300 pounds, and it was a hot day.